+++ src/reducers/books.js
@@ -1,11 +1,11 @@
 import { CREATE_BOOK, REMOVE_BOOK } from '../actions';
 
-const booksReducer = (state = [], action) => {
-  switch (action.type) {
+const booksReducer = (state = [], {type, payload}) => {
+  switch (type) {
     case CREATE_BOOK:
-      return [...state, action.payload];
+      return [...state, payload];
     case REMOVE_BOOK:
-      return state.filter(book => book.id !== action.payload);
+      return state.filter(book => book.id !== payload);
     default:
       return state;
   }
